youtube_shorts/requirements.txt
ollama>=0.1.14
gtts>=2.3.1
pydub>=0.25.1
pillow>=9.2.0
google-auth>=2.15.0
google-auth-oauthlib>=0.5.3
google-api-python-client>=2.70.0
pyyaml>=6.0
requests>=2.28.1
